refactor: merge canopy into core
