+++ crates/canopy-core/Cargo.toml
@@ -16,3 +16,7 @@ rhai = { version = "1.22.2", features = ["internals"] }
 regex = "1.11.1"
 termcolor = "1.4.1"
 geom = { path = "../geom" }
+comfy-table = "7.1.4"
+color-backtrace = "0.7.0"
+scoped-tls = "1.0.1"
+expectrl = "0.7.1"
