+++ crates/canopy-core/src/backend/dummy.rs
@@ -1,4 +1,4 @@
-use canopy_core::{Result, geom::Point, render::RenderBackend, style::Style};
+use crate::{Result, geom::Point, render::RenderBackend, style::Style};
 
 /// A dummy render backend that discards all output.
 /// This is useful for tests where we want to inspect the TermBuf directly.
