+++ crates/canopy-core/src/backend/test.rs
@@ -1,9 +1,7 @@
 use crate::Canopy;
-#[cfg(test)]
-use canopy_core::geom::Expanse;
-use canopy_core::{
+use crate::{
     Node, Result,
-    geom::Point,
+    geom::{Expanse, Point},
     render::RenderBackend,
     style::{Style, StyleManager},
 };
