+++ crates/canopy-core/src/tutils/harness.rs
@@ -3,7 +3,7 @@ use crate::{
     Canopy, Loader,
     backend::{dummy::DummyBackend, test::TestRender},
 };
-use canopy_core::{Node, Result, TermBuf, event::key, geom::Expanse};
+use crate::{Node, Result, TermBuf, event::key, geom::Expanse};
 
 /// Run a function on our standard dummy app built from [`ttree`]. This helper
 /// is used extensively in unit tests across the codebase.
