+++ crates/canopy/Cargo.toml
@@ -5,25 +5,10 @@ version = "0.0.1"
 edition = "2024"
 
 [dependencies]
-crossterm = "0.29.0"
-pad = "0.1.6"
-textwrap = "0.16.2"
-color-backtrace = "0.7.0"
-scopeguard = "1.2.0"
-
 canopy-core = { path = "../canopy-core" }
 canopy-widgets = { path = "../canopy-widgets" }
-canopy-derive = { path = "../canopy-derive" }
-thiserror = "2.0.12"
-bitvec = "1.0.1"
 tracing-subscriber = "0.3.19"
 tracing = "0.1.41"
-comfy-table = "7.1.4"
-convert_case = "0.8.0"
-regex = "1.11.1"
-rhai = { version = "1.22.2", features = ["internals"] }
-scoped-tls = "1.0.1"
-expectrl = "0.7.1"
 
 [dev-dependencies]
 clap = { version = "4.5.40", features = ["derive"] }
