+++ crates/canopy/src/lib.rs
@@ -4,25 +4,18 @@
 // Re-export everything from canopy-core
 pub use canopy_core::*;
 
-// Modules that remain in canopy
-mod canopy;
-mod focus;
-mod inputmap;
-mod poll;
-mod root;
-
-pub mod backend;
-mod binder;
 pub mod inspector;
-pub mod script;
+mod root;
 // Re-export widgets from canopy-widgets
 pub use canopy_widgets as widgets;
 
 // Re-export canopy-specific types
-pub use crate::canopy::{Canopy, Loader};
-pub use binder::*;
+pub use canopy_core::{
+    Binder, Canopy, DefaultBindings, FocusableNode, Input, InputMap, InputMode, Loader, Poller,
+    collect_focusable_nodes, find_focus_target, find_focused_node,
+};
 pub use root::*;
 
 // Hide the test utils from docs. We need to expose it for integration tests, but it's not for external use.
 #[doc(hidden)]
-pub mod tutils;
+pub use canopy_core::tutils;
